refactor: remove usage and backups
